Move number list into own file
